Support deep sleep for the power saving
Due to the system specification, it's impossible to enter the deep
sleeping mode for the power saving as long as USB is attached.
Anymore SPRESENSE board works as USB bus powered device, and can't
support for deep sleep which means the lowest power consumption. As
the workaround for this restriction, this configuration forces to be
disable the USB function in deep sleeping.

refs #6801