Fix both edge detection of gpio interrupt
When gpio interrupt has the polarity of both edge, it doesn't work well
if the short-pulse signal is detected. This problem is caused by the
restriction of the gpio hardware specification. To avoid this, add a
software workaround which realizes the detection of both edge by using
the inversion of interrupt mask.

refs #8138