examples/awsiot_gnsslogger: Fix the issue that may not connect to AWSIoT
The example may not connect to AWS IoT.
Because available time period of cert file, which is generated by each
"Things" on AWS IoT, is started when the file is generated, the system
time on the board needs to be set current time when connecting to AWS
IoT server.
So this example is modified to set the system time from the gnss
measurement result before connecting.