examples/awsiot_gnsslogger: Fix overflow ini file max line
This example provides template of ini file as examples/awsiot_gnsslogger/spif/aws_iot.ini.
In this file, the comment line is over 80 charactors.
Current default max line number of inih library is 80, so it will overflow.
On this commit will change maximum number of the line from 80 to 128 on the defconfig,
and fixed the comment line to be less than 80 charactors in aws_iot.ini file just in case.