sdk/modules/mbedtls_stub: Address additional TLS command
The alt1250 driver has mbedTLS capability, but it is not
supported by the standard driver code.
This commit provides a handler to use the mbedTLS feature
of the driver.
The alt1250 provides a handler function as a weak symbol,
so overriding the weak handler by your own handler makes
enable the mbedTLS feature.

At the same time, the "--whole-archive" option has been
added to LDFLAGS. This option force the linker to search
all archives for symbols to replace the weak symbols.